Check entity is null to avoid exception when deleting entity

+++ Solution/Scalesoft.Localization.Database.NHibernate/UnitOfWork/StaticTextUoW.cs
@@ -88,8 +88,11 @@ namespace Scalesoft.Localization.Database.NHibernate.UnitOfWork
                     name, cultureName, dictionaryScopeName
                 );
 
-                staticTextRepository.Delete(staticTextEntity);
-                session.Flush();
+                if (staticTextEntity != null)
+                {
+                    staticTextRepository.Delete(staticTextEntity);
+                    session.Flush();
+                }
             }
         }
